Flytt ut økonomiutils
