--- src/main/resources/test.json
@@ -1,42 +1,0 @@
-{
-  "kodeEndring": "UEND",
-  "fagSystem": "BA",
-  "saksnummer": "999952",
-  "aktoer": "12345678910",
-  "saksbehandlerId": "F_Z994955.E_Z994955@trygdeetaten.no",
-  "avstemmingTidspunkt": "2020-07-01T13:43:15.577437",
-  "utbetalingsperiode": [
-    {
-      "erEndringPåEksisterendePeriode": true,
-      "opphør": {
-        "opphørDatoFom": "2020-06-01"
-      },
-      "periodeId": 999951000,
-      "forrigePeriodeId": null,
-      "datoForVedtak": "2020-07-01",
-      "klassifisering": "BATR",
-      "vedtakdatoFom": "2020-06-01",
-      "vedtakdatoTom": "2037-04-30",
-      "sats": 1054,
-      "satsType": "MND",
-      "utbetalesTil": "12345678910",
-      "behandlingId": 999952
-    },
-    {
-      "erEndringPåEksisterendePeriode": true,
-      "opphør": {
-        "opphørDatoFom": "2018-12-01"
-      },
-      "periodeId": 999951002,
-      "forrigePeriodeId": 999951001,
-      "datoForVedtak": "2020-07-01",
-      "klassifisering": "BATR",
-      "vedtakdatoFom": "2019-03-01",
-      "vedtakdatoTom": "2036-04-30",
-      "sats": 1054,
-      "satsType": "MND",
-      "utbetalesTil": "12345678910",
-      "behandlingId": 999952
-    }
-  ]
-}
